+++ c/gomlx/xlabuilder/gen_op_types.h
@@ -89,6 +89,7 @@ enum OpType {
   LessThanTotalOrderOp,
   DynamicSliceOp,
   DynamicUpdateSliceOp,
+  ErfOp,
 };
 
 #ifdef __cplusplus