+++ xlabuilder/gen_op_types.go
@@ -86,4 +86,5 @@ const (
 	LessThanTotalOrderOp
 	DynamicSliceOp
 	DynamicUpdateSliceOp
+	ErfOp
 )